# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageErrorEnvelope'
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageErrorEnvelope'
  # ─── ENTITIES ───────────────────────────────────────────────
...
  /wards:
    get:
      operationId: listWards
      tags: [wards]
      summary: List wards
      responses:
        "200":
          description: Wards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ward"
    post:
      operationId: createWard
      tags: [wards]
      summary: Create ward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWardBody"
      responses:
        "201":
          description: Created ward
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ward"
...
  /beds:
    get:
      operationId: listBeds
      tags: [wards]
      summary: List beds
      parameters:
        - name: wardId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Beds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bed"
    post:
      operationId: createBed
      tags: [wards]
      summary: Create bed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBedBody"
      responses:
        "201":
          description: Created bed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bed"

  # ─── PHARMACY ───────────────────────────────────────────────
...
components:
  schemas:
...
    Ward:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        totalBeds:
          type: integer
        availableBeds:
          type: integer
        ratePerDay:
          type: number
      required: [id, name, type, totalBeds]
    CreateWardBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        ratePerDay:
          type: number
      required: [name, type]
    Bed:
      type: object
      properties:
        id:
          type: integer
        wardId:
          type: integer
        wardName:
          type: string
        bedNo:
          type: string
        status:
          type: string
        patientName:
          type: string
      required: [id, wardId, bedNo, status]
    CreateBedBody:
      type: object
      properties:
        wardId:
          type: integer
        bedNo:
          type: string
      required: [wardId, bedNo]
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    StorageErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }